+++ react-components/src/components/Form/Select/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from 'react'
+import React, { useState, useRef, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { Input } from '@src/components/Form/Input'
 import { useOnOutsideClick } from '@src/components/hooks/useOnOutsideClick'
@@ -28,6 +28,10 @@ export const Select = ({
   const liRef = useRef([])
   const [element] = useOnOutsideClick(() => setIsOpen(false))
   useNoScroll(isOpen)
+
+  useEffect(() => {
+    setInput(selected)
+  }, [selected])
 
   const selectOption = (item) => {
     setInput(item.label)