+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 
 ### Implemented enhancements
 
+- GP2-1363 - Drop custom Django view serving /markets/
+- GP2-1340 - Analytics - product name relabelled
+- GP2-1339 - Analytics - search again
+- GP2-1260 - Compare markets - tooltips
+- GP2-1132 - Product classifier - surface definitions
 - GP2-341 - Adaptation language info and stats update
 - GP2-1270 - income data for economy tab
 - GP2-1318 - Compare markets - refactor and loading twinkle
 - GP2-1265 - Compare markets - society data attribution
 - GP2-1258 - Compare markets - integrate society data
+- GP2-1284 - Port over CountryGuidePage from BAU, including base template, header, footer, etc
 - GP2-1270 - income data for economy tab
 - GP2-1259 - Styling for compare markets datatable from great-styles
 - GP2-1309 - Remove tooltips
@@ -36,9 +42,19 @@
 - GP2-1111 - EP responsive marketing approach changes
 - GP2-1029 - Population data sources attributions and links
 - GP2-974 - target market research page
+- GP2-285 - cost and price be
+- GP2-1323 - enable ep sections
+- GP2-1321 - add TM  learning component
+- GP2-1306 - learning data
+- GP2-1168 - export plan lessons
 
 ### Fixed bugs
 
+- GP2-1365 - Add bulleted list styles to rich text block
+- GP2-1364 - Fix extra spacing for first element in rich text block
+- GP2-1366 - minor fix of help text for rich-text areas
+- GP2-1367 - minor fix of link field label
+- No ticket - Next lesson button showing blue underline
 - No ticket - Remove tooltip from languages panel on adaptation for tarket market
 - No ticket - Added missing population url
 - GP2-1310 - Data not loading for if made more than one country is compared